% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.simulation.R
\name{analysis.simulation}
\alias{analysis.simulation}
\title{Simulation analysis}
\usage{
analysis.simulation(x)
}
\arguments{
\item{x}{A simulation vector}
}
\description{
Creates a simple PDF analysis for a simulation vector. PDF is saved in current working directory
}
\examples{

x <- MC_pi
class(x) <- 'simulation'
analysis(x)

}
