% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARsim.R
\name{ARsim}
\alias{ARsim}
\title{ARsim}
\usage{
ARsim(f, C, N, exponential = TRUE, rate = 1, propdensity = NA,
  inveresepropdist = NA)
}
\arguments{
\item{f}{target density}

\item{C}{constant such that f(x) <= C*g(x)}

\item{N}{number of samples}

\item{exponential}{indicates if we want to use the exponential density as proposal}

\item{rate}{corresponding rate}

\item{propdensity}{alternative proposal if needed}

\item{inveresepropdist}{corresponding inverse distribution function}
}
\value{
returns a list containing the value of the integral, the 0.95-confidence interval, number of simulations and the function f
}
\description{
The function ARsim() draws from the random variable, which is specified by the density f, using acceptance rejection sampling. 
The default proposal density is the exponential density with rate 1. If you want to use a different density you have to set the argument "exponential" to FALSE and specify your new proposal density and the corresponding inverse distribution function.
}
\examples{
C <- 1.6
lambda <- 0.5
f <- function(x) {x*exp(-x)}

ARsimulation <- ARsim(f, C, 100000, rate = lambda)

}
