% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC_integration.R
\name{int}
\alias{int}
\title{int}
\usage{
int(f, a, b, n = 10^5)
}
\arguments{
\item{f}{the integrand}

\item{a}{lower bound of integration}

\item{b}{upper bound of integration}

\item{n}{number of simulation, per default 10^5}
}
\value{
returns a list containing the value of the integral, the 0.95-confidence interval, number of simulations and the function f
}
\description{
The function int() performs a Monte Carlo integration of the function f.
}
\examples{
f <- function(x){x^2}
MCint <- int(f, -2, 2)
MCint$'Value of integral'

}
