% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invsample.R
\name{inv_sample}
\alias{inv_sample}
\title{inv_sample}
\usage{
inv_sample(Finv, N = 1, F_distribution = NA)
}
\arguments{
\item{Finv}{inverse distribution function}

\item{N}{number of samples}

\item{F_distribution}{distribution function, optional for more sophisticated plots}
}
\value{
returns a list containing the samples, the number of simulations and (optional) the distribution function.
}
\description{
The function inv_sample() draws from the random variable, which is specified by the distribution F, using inverse sampling.
}
\examples{
Finv <- function(x){-log(1-x)}
F_dist <- function(x){1-exp(-x)}

inv_samples <- inv_sample(Finv, N = 10000)
inv_samples_withF <- inv_sample(Finv, N = 10000, F_distribution = F_dist)
inv_5samples <- inv_sample(Finv, N = 5)

}
