% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MC_pi}
\alias{MC_pi}
\title{The result of a Crude Monte Carlo simulation of pi}
\format{A column vector with 1000000 rows

\describe{
  \item{MC_pi}{Pi simulated with pairs of 2 independet (0,1) uniforms, and the evaluating the ratio that falls in the unit circle}
}}
\usage{
MC_pi
}
\description{
A vector containing simulation of pi
}
\keyword{datasets}
