% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.AR.R
\name{plot.AR}
\alias{plot.AR}
\title{plot.AR}
\usage{
\method{plot}{AR}(ARsample)
}
\arguments{
\item{ARsample}{samples created by acceptance rejection}
}
\description{
The function plot.ARs() plots the target vs. the proposal density and the empirical vs. the theoretical density of the samples.
}
\examples{
C <- 1.6
lambda <- 0.5
f <- function(x) {x*exp(-x)}

ARsimulation <- ARsim(f, C, 100000, rate = lambda)
plot(ARsimulation)
}
