% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.copula.R
\name{plot.copula}
\alias{plot.copula}
\title{plot.copula}
\usage{
\method{plot}{copula}(data)
}
\arguments{
\item{data}{2 dimensional matrix}
}
\value{
plot of copula
}
\description{
plot.copula plots a 2 dimensional copula, using empirical CDF.
}
\examples{
N <- 1000
data <- rmult_norm(10000,rho=0.7)
class(data) <- 'copula'
plot(data)

}
