% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.invsampling.R
\name{plot.invsample}
\alias{plot.invsample}
\title{plot.invsample}
\usage{
\method{plot}{invsample}(inv_sample)
}
\arguments{
\item{inv_sample}{}
}
\value{
returns a plot of the empirical distribution function
}
\description{
The function plot.invsample() plots the empirical distribution function of the sampled values
}
\examples{
Finv <- function(x){-log(1-x)}
F_dist <- function(x){1-exp(-x)}

inv_samples <- inv_sample(Finv, N = 10000)
inv_samples_withF <- inv_sample(Finv, N = 10000, F_distribution = F_dist)

plot(inv_samples)
plot(inv_samples_withF)

}
