% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.MC_integral.R
\name{plot.MC_integral}
\alias{plot.MC_integral}
\title{plot.MC_integral}
\usage{
\method{plot}{MC_integral}(MC, ggplot = FALSE)
}
\arguments{
\item{MC}{output of a Monte Carlo integration}

\item{ggplot}{indicates if you want to use ggplot; please note that ggplot has to be installed in this case}
}
\value{

}
\description{
The function plot.MC_integral plots the integrand of a corresponding MC integration
}
\examples{
f <- function(x){x^2}
MCint <- int(f, -2, 2)
plot(MCint)

}
