% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.simulation.R
\name{plot.simulation}
\alias{plot.simulation}
\title{plot.simulation}
\usage{
\method{plot}{simulation}(x, ggplot = FALSE)
}
\arguments{
\item{x}{A result vector from a simulation algorithm}

\item{ggplot}{A boolean variable, determining if to use ggplot, directly from function call, rather than using control input}
}
\value{
A plot of the summary of the input vector, containing mean, variance, relative error and confidence interval
}
\description{
plot.simulation
}
\examples{
x <- MC_pi
plot(x)

}
