% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_analysis.R
\name{simulation_analysis}
\alias{simulation_analysis}
\title{Simulation analysis}
\usage{
simulation_analysis(x)
}
\arguments{
\item{x}{A simulation vector}
}
\description{
Creates a simple PDF analysis for a simulation vector. PDF is saved in current working directory
}
\examples{

x <- MC_pi
class(x) <- 'simulation'
analysis(x)

}
