% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.copula.R
\name{plot.copula}
\alias{plot.copula}
\title{plot.copula}
\usage{
\method{plot}{copula}(data, ggplot = FALSE)
}
\arguments{
\item{data}{2 dimensional matrix}
}
\value{
plot of copula
}
\description{
plot.copula plots a 2 dimensional copula, using empirical CDF.
}
\examples{
#Example using bivariate normal
N <- 1000
data <- rbi_norm(10000,rho=0.7)
class(data) <- 'copula'
plot(data)


#Example using 1-Fréchet
N <- 10000
R <- -1/log(runif(N))
U <- runif(N)
X <- R*U
Y <- R*(1-U)
data <- matrix(c(X,Y),ncol=2)
class(data) <- 'copula'
plot(data, ggplot=T)


}
