% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.simulation.R
\name{summary.simulation}
\alias{summary.simulation}
\title{summary.simulation}
\usage{
\method{summary}{simulation}(x)
}
\arguments{
\item{x}{A result vector from a simulation algorithm}
}
\description{
summary.simulation
}
